db implementation

+++ app/src/main/java/com/pranjal/codersclock/models/CodeforcesContest.kt
@@ -1,12 +1,12 @@
 package com.pranjal.codersclock.models
 
 data class CodeforcesContest(
-    val durationSeconds: Int,
+    val durationSeconds: Long,
     val frozen: Boolean,
     val id: Int,
     val name: String,
     val phase: String,
-    val relativeTimeSeconds: Int,
-    val startTimeSeconds: Int,
+    val relativeTimeSeconds: Long,
+    val startTimeSeconds: Long,
     val type: String
 )
